LazyLogging: Use lazy log interpolation everywhere
One of the main benefits of the python logging framework is the ability to
defer the string operations to construct the full log message until the
message is formatted. This allows more granular logging without the cost of
many extra string operations since those operations will not be performed
if the log level is disabled.

This does require use of the more archaic percent-encoding style of string
formatting, but the performance gain (and general best practice) is worth
the gorpiness!

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>