input added all at once treated differently than one at a time
There is no spec for rouge, so in a sense there can't really be bugs, but it is very unintuitive that a list of inputs, when added all at once, are treated differently than when they are added one at a time, in the same order that the rouge code processes them.

This  is because the rouge code tests inputs against ALL inputs, whether or not they are added.

My earlier PR (#69) avoided that problem, but it has not been accepted.